bib: move helpers into util.go
Tiny cleanup/refactor to move some code out of {main,images}.go
so that those places are less cluttered.
