+++ bib/cmd/bootc-image-builder/util.go
@@ -1,0 +1,76 @@
+package main
+
+import (
+	cryptorand "crypto/rand"
+	"fmt"
+	"math"
+	"math/big"
+	"math/rand"
+	"os"
+	"path/filepath"
+	"strconv"
+	"strings"
+)
+
+// canChownInPath checks if the ownership of files can be set in a given path.
+func canChownInPath(path string) (bool, error) {
+	info, err := os.Stat(path)
+	if err != nil {
+		return false, err
+	}
+	if !info.IsDir() {
+		return false, fmt.Errorf("%s is not a directory", path)
+	}
+
+	checkFile, err := os.CreateTemp(path, ".writecheck")
+	if err != nil {
+		return false, err
+	}
+	defer func() {
+		if err := os.Remove(checkFile.Name()); err != nil {
+			// print the error message for info but don't error out
+			fmt.Fprintf(os.Stderr, "error deleting %s: %s\n", checkFile.Name(), err.Error())
+		}
+	}()
+	return checkFile.Chown(osGetuid(), osGetgid()) == nil, nil
+}
+
+func chownR(path string, chown string) error {
+	if chown == "" {
+		return nil
+	}
+	errFmt := "cannot parse chown: %v"
+
+	var gid int
+	uidS, gidS, _ := strings.Cut(chown, ":")
+	uid, err := strconv.Atoi(uidS)
+	if err != nil {
+		return fmt.Errorf(errFmt, err)
+	}
+	if gidS != "" {
+		gid, err = strconv.Atoi(gidS)
+		if err != nil {
+			return fmt.Errorf(errFmt, err)
+		}
+	} else {
+		gid = osGetgid()
+	}
+
+	return filepath.Walk(path, func(name string, info os.FileInfo, err error) error {
+		if err == nil {
+			err = os.Chown(name, uid, gid)
+		}
+		return err
+	})
+}
+
+func createRand() *rand.Rand {
+	seed, err := cryptorand.Int(cryptorand.Reader, big.NewInt(math.MaxInt64))
+	if err != nil {
+		panic("Cannot generate an RNG seed.")
+	}
+
+	// math/rand is good enough in this case
+	/* #nosec G404 */
+	return rand.New(rand.NewSource(seed.Int64()))
+}
